Chorus, vibrato AM-index, AM Index, ellipsoid, pitch manager, traces
Native chorus: multi-voice delay with shared vibrato (Top/Side Morph), HP/LP on wet, no feedback, MLR I/O.

Vibrato Generator: wavetable sine, Top Morph AM, Side Morph phase, Speed/Phase defaults.

AM Index (Bias + Amplitude). Ellipsoid Oscillator 2D + Limit AA. Pitch Manager simultaneous Inc/f/pitch and octave live-param map.

Trace wires: feedback U under modules, forward quantized HVH. Range spawn, controller Display/Folder/Control persist, invalid patch wires disconnect, Inc labels.

+++ native_modules/amp_curve/amp_curve.cpp
@@ -11,6 +11,8 @@
 //   Lin: Out = x
 //   Exp: Out = 0 at x≤0, else 10^(k·(x−1))  (unity at 1, ~−100 dB near 0)
 
+#include <cstdint>
+
 #include "../sandbox_native_maths/sandbox_native_maths.h"
 
 namespace {
@@ -110,14 +112,14 @@ extern "C" void soemdsp_amp_curve_process_block(int handle, int frameCount) {
   }
 }
 
-extern "C" int soemdsp_amp_curve_block_input_ptr(int handle) {
+extern "C" intptr_t soemdsp_amp_curve_block_input_ptr(int handle) {
   if (handle < 1 || handle > kMaxInstances) return 0;
-  return reinterpret_cast<int>(gPool[handle - 1].blockIn);
+  return reinterpret_cast<intptr_t>(gPool[handle - 1].blockIn);
 }
 
-extern "C" int soemdsp_amp_curve_block_output_ptr(int handle) {
+extern "C" intptr_t soemdsp_amp_curve_block_output_ptr(int handle) {
   if (handle < 1 || handle > kMaxInstances) return 0;
-  return reinterpret_cast<int>(gPool[handle - 1].blockOut);
+  return reinterpret_cast<intptr_t>(gPool[handle - 1].blockOut);
 }
 
 extern "C" int soemdsp_amp_curve_max_block_frames() {
